Tentative acceleration clamping implemented. Untested.

+++ Trunk/LRT12/ActionData/IMUData.h
@@ -4,7 +4,7 @@
 struct IMU_Data
 {
 	double roll, pitch, yaw;
-	int16_t accel_x, accel_y, accel_z;
+	double accel_x, accel_y, accel_z;
 	int16_t gyro_x, gyro_y, gyro_z;
 };
